Implementing domain and infrastructure layers
