+++ ClassroomManagement.Domain/Entities/BaseEntiy.cs
@@ -1,0 +1,6 @@
+﻿namespace ClassroomManagement.Domain.Entities;
+
+public abstract class BaseEntiy
+{
+    public int Id { get; set; }
+}
