--- ClassroomManagement/Models/ErrorViewModel.cs
@@ -1,9 +1,0 @@
-namespace ClassroomManagement.Models
-{
-    public class ErrorViewModel
-    {
-        public string? RequestId { get; set; }
-
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
-}
