--- ClassroomManagement/Models/ViewModel/AverageScoreGroup.cs
@@ -1,13 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace ClassroomManagement.Models.ViewModel
-{
-    public class AverageScoreGroup
-    {
-        public int Id { get; set; }
-        public string? Name { get; set; }
-
-        [DisplayFormat(DataFormatString = "{0:0.00}")]
-        public double Average { get; set; }
-    }
-}
