--- ClassroomManagement/Models/ViewModel/ProfessoresViewModel.cs
@@ -1,9 +1,0 @@
-﻿namespace ClassroomManagement.Models.ViewModel
-{
-    public class ProfessoresViewModel
-    {
-        public int Id { get; set; }
-        public string? ProfessorName { get; set; }
-        public string? Subject { get; set; }
-    }
-}
