+++ ClassroomManagement/ViewModels/ErrorViewModel.cs
@@ -1,0 +1,8 @@
+namespace ClassroomManagement.Web.ViewModels;
+
+public class ErrorViewModel
+{
+    public string? RequestId { get; set; }
+
+    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+}
