+++ ClassroomManagement/ViewModels/ProfessoresViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace ClassroomManagement.Web.ViewModels;
+
+public class ProfessoresViewModel
+{
+    public int Id { get; set; }
+    public string? ProfessorName { get; set; }
+    public string? Subject { get; set; }
+}
