+++ ClassroomManagement/Views/Professor/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model ClassroomManagement.Models.ViewModel.ProfessorExam
+﻿@model ProfessorExam
 
 @{
     ViewData["Title"] = "Detalhes";
