+++ ClassroomManagement/Views/Professor/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<ClassroomManagement.Models.ViewModel.ProfessoresViewModel>
+﻿@model IEnumerable<ProfessoresViewModel>
 
 @{
     ViewData["title"] = "Professores";
