+++ ClassroomManagement/Views/Student/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model ClassroomManagement.Models.Student
+﻿@model Student
 
 @{
     ViewData["Title"] = "Adicionar";
