+++ ClassroomManagement/Views/Student/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@model ClassroomManagement.Models.Student
+﻿@model Student
 
 @{
     ViewData["Title"] = "Deletar";
