+++ ClassroomManagement/Views/Student/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<ClassroomManagement.Models.ViewModel.AverageScoreGroup>
+﻿@model IEnumerable<AverageScoreGroup>
 
 @{
     ViewData["Title"] = "Tabela";
