+++ ClassroomManagement/Views/StudentViewModel/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@model ClassroomManagement.Models.ViewModel.StudentViewModel
+﻿@model StudentViewModel
 
 @{
     ViewData["title"] = "Deletar";
