+++ ClassroomManagement/Views/Subjects/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model ClassroomManagement.Models.Subject
+﻿@model Subject
 
 @{
     ViewData["title"] = "Criar";
