+++ ClassroomManagement/Views/Subjects/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@model ClassroomManagement.Models.Subject
+﻿@model Subject
 
 @{
     ViewData["title"] = "Deletar";
