+++ ClassroomManagement/Views/Subjects/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<ClassroomManagement.Models.Subject>
+﻿@model IEnumerable<Subject>
 
 @{
     ViewData["Title"] = "Materia";
