+++ ClassroomManagement/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-﻿@using ClassroomManagement
-@using ClassroomManagement.Models
+﻿@using ClassroomManagement.Web.ViewModels
+@using ClassroomManagement.Domain.Entities
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
